DCMIP ICs implemented
This version passes the TestStateConstructors portion of the State test with analytic_init() but fails the testStateInterpolation portion.  This appears to be an issue with the interp_t() routine in fv3jedi_fields_mod.F90; the test_fv3jedi_state fails for the same reason.  Since fixing this was not the point of this feature branch (I'm not sure who else might be working on it), I'm ready to merge this back into develop.