Update fv3-jedi's atlas interface (#1087)
* Remove toFieldSetAD

(This is an inconsistent state: oops is at this point expecting a
functional atlas.haloExchange, which is not yet provided. It will
be in a following commit and all squashed together...)

* Copy owned data only in to/fromFieldSet; use atlas halo exchanges to update halos

(This is an inconsistent state: oops is at this point expecting a
functional atlas.haloExchange, which is not yet provided. It will
be in a following commit and all squashed together...)

* Rename two FunctionSpaces

* Upgrade FunctionSpace to NodeColumns by importing the fv3 cubed-sphere connectivity

* Pass MPI comm to atlas

* Clean geometry includes

* Update test refs; new halo changes interp results

* Fix setup of bump interpolator

* Require recent tag of atlas

* Fix ref

* Review comment

* Trigger CI